Move `functionobject.cc` to `api/function.cc`.
