+++ src/api/function.h
@@ -1,5 +1,5 @@
-#ifndef TEMPEARLY_FUNCTIONOBJECT_H_GUARD
-#define TEMPEARLY_FUNCTIONOBJECT_H_GUARD
+#ifndef TEMPEARLY_API_FUNCTION_H_GUARD
+#define TEMPEARLY_API_FUNCTION_H_GUARD
 
 #include "object.h"
 
@@ -23,24 +23,6 @@ namespace tempearly
          * Default destructor.
          */
         virtual ~FunctionObject();
-
-        /**
-         * Constructs method.
-         *
-         * \param interpreter Script interpreter
-         * \param cls         Class which declares the method
-         * \param name        Name of the method
-         * \param arity       Method arity, e.g. how many arguments the method
-         *                    takes
-         * \param callback    Callback function which is invoked when the method
-         *                    is invoked
-         * \return            Value which is the constructed method
-         */
-        static Value NewMethod(const Handle<Interpreter>& interpreter,
-                               const Handle<Class>& cls,
-                               const String& name,
-                               int arity,
-                               Callback callback);
 
         /**
          * Invokes the function.
@@ -74,4 +56,4 @@ namespace tempearly
     };
 }
 
-#endif /* !TEMPEARLY_FUNCTIONOBJECT_H_GUARD */
+#endif /* !TEMPEARLY_API_FUNCTION_H_GUARD */
